pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "aa_relays"
dynamic = ["version", "description" ]
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.8"
authors = [
    { name = "Joel Falknau", email = "joel.falknau@gmail.com" },
]
keywords = [
    "allianceauth",
    "eveonline",
]
classifiers = [
    "Environment :: Web Environment",
    "Framework :: Celery",
    "Framework :: Django",
    "Framework :: Django :: 4.0",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: Implementation :: CPython",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content"
]
dependencies = [
    "allianceauth>=3.0.0,<5.0.0",
    "django-eveuniverse",
    "allianceauth-discordbot",
    "django-solo>=2.0.0,<3.0.0",
    "py-cord>=2.0.0,<3.0.0",
    "googletrans==4.0.0-rc1",
    "selfcord-py@git+https://github.com/dolfies/discord.py-self@renamed#egg=selfcord.py",
]

[project.urls]
Homepage = "https://gitlab.com/tactical-supremacy/aa-relays/"
Source = "https://gitlab.com/tactical-supremacy/aa-relays/"
Tracker = "https://gitlab.com/tactical-supremacy/aa-relays/-/issues"

[tool.flit.module]
name = "relays"

[tool.isort]
profile = "django"
sections = [
    "FUTURE",
    "STDLIB",
    "THIRDPARTY",
    "DJANGO",
    "ALLIANCEAUTH",
    "FIRSTPARTY",
    "LOCALFOLDER"
]
known_allianceauth = ["allianceauth", "esi"]
known_django = ["django"]
skip_gitignore = true

[tool.flake8]
exclude = [".git", "*migrations*", ".tox", "dist", "htmlcov"]
max-line-length = 119
select = ["C", "E", "F", "W", "B", "B950"]
ignore = ['E203', 'E231', 'E501', 'W503', 'W291', 'W293']

[tool.djlint]
max_attribute_length=119
max_line_length=119
max_blank_lines=1
